KAN-48 fix: add copmponents for SidebarMenu, logaut Modal, deleted logic, deleted ClientRootLayot
