--- src/app/ClientRootLayout.tsx
@@ -1,165 +1,0 @@
-'use client'
-
-import { useLogoutMutation, useMeQuery } from '@/features/auth/api'
-import { Button } from '@/shared/components/Button'
-import { Loader } from '@/shared/components/Loader'
-import { Modal, ModalBody, ModalClose, ModalTitle } from '@/shared/components/Modal'
-import { Typography } from '@/shared/components/Typography'
-import { PublicRoutes } from '@/shared/enums'
-import {
-   CreateIcon,
-   CreateOutlineIcon,
-   CrossIcon,
-   FavoriteIcon,
-   FavoriteOutlineIcon,
-   HomeIcon,
-   HomeOutlineIcon,
-   LogoutIcon,
-   MessageIcon,
-   MessageOutlineIcon,
-   PersonIcon,
-   PersonOutlineIcon,
-   SearchIcon,
-   StatisticIcon,
-} from '@/shared/icons'
-import { cn } from '@/shared/lib'
-import { Sidebar, SidebarItem } from '@/widgets/Sidebar'
-import { usePathname, useRouter } from 'next/navigation'
-import { ReactNode, useState } from 'react'
-
-const PUBLIC_ROUTES = Object.values(PublicRoutes) as string[]
-
-export default function ClientRootLayout({ children }: { children: ReactNode }) {
-   const [isLogoutModalOpen, setIsLogoutModalOpen] = useState(false)
-
-   const token = typeof window !== 'undefined' ? localStorage.getItem('accessToken') : null
-
-   const { data: user, isLoading } = useMeQuery(undefined, {
-      skip: !token,
-   })
-
-   const [logout] = useLogoutMutation()
-   const router = useRouter()
-
-   const handleLogoutClick = () => {
-      setIsLogoutModalOpen(true)
-   }
-
-   const handleCloseModal = () => {
-      setIsLogoutModalOpen(false)
-   }
-
-   const handleConfirmLogout = async () => {
-      await logout()
-      setIsLogoutModalOpen(false)
-      router.push(PublicRoutes.signIn)
-   }
-
-   const pathname = usePathname()
-   const isPublic = PUBLIC_ROUTES.includes(pathname)
-
-   if (isLoading) {
-      return (
-         <div className="flex h-screen w-full items-center justify-center">
-            <Loader size="8rem" />
-         </div>
-      )
-   }
-
-   return (
-      <div
-         className={cn(
-            'mx-auto flex w-full max-w-[1280px] flex-col items-center px-[60px]',
-            isPublic ? 'items-center px-[0px]' : 'items-start'
-         )}
-      >
-         {!isPublic && (
-            <>
-               <Sidebar className="top-[60px] min-w-[180px] pl-[0px]">
-                  <SidebarItem
-                     id="1"
-                     name="Feed"
-                     icon={HomeOutlineIcon}
-                     activeIcon={HomeIcon}
-                     path="/feed"
-                  />
-                  <SidebarItem
-                     id="2"
-                     name="Create"
-                     icon={CreateOutlineIcon}
-                     activeIcon={CreateIcon}
-                     path="/create"
-                  />
-                  <SidebarItem
-                     id="3"
-                     name="My Profile"
-                     icon={PersonOutlineIcon}
-                     activeIcon={PersonIcon}
-                     path="/profile"
-                  />
-                  <SidebarItem
-                     id="4"
-                     name="Messenger"
-                     icon={MessageOutlineIcon}
-                     activeIcon={MessageIcon}
-                     path="/messenger"
-                  />
-                  <SidebarItem
-                     id="5"
-                     name="Search"
-                     icon={SearchIcon}
-                     path="/search"
-                     className="mb-15"
-                  />
-                  <SidebarItem id="6" name="Statistics" icon={StatisticIcon} path="/statistics" />
-                  <SidebarItem
-                     id="7"
-                     name="Favorites"
-                     icon={FavoriteOutlineIcon}
-                     activeIcon={FavoriteIcon}
-                     path="/favorites"
-                     className="mb-45"
-                  />
-                  <SidebarItem
-                     id="8"
-                     name="Log Out"
-                     icon={LogoutIcon}
-                     path="/signup"
-                     onClick={handleLogoutClick}
-                  />
-               </Sidebar>
-
-               <Modal open={isLogoutModalOpen} onOpenChange={setIsLogoutModalOpen}>
-                  <ModalTitle className={'flex items-center justify-between'}>
-                     <Typography variant={'h1'}>Log Out</Typography>
-                     <ModalClose asChild>
-                        <CrossIcon />
-                     </ModalClose>
-                  </ModalTitle>
-
-                  <hr className={'text-dark-100 h-[1px]'} />
-
-                  <ModalBody className="flex max-w-[378px] flex-col gap-4 px-6 py-3">
-                     <Typography>
-                        Are you really want to log out of your account {user?.email}?
-                     </Typography>
-                     <span className="flex gap-2 self-end">
-                        <Button
-                           onClick={handleCloseModal}
-                           variant="outlined"
-                           className={'h-[36px] w-[96px]'}
-                        >
-                           No
-                        </Button>
-                        <Button onClick={handleConfirmLogout} className={'h-[36px] w-[96px]'}>
-                           Yes
-                        </Button>
-                     </span>
-                  </ModalBody>
-               </Modal>
-            </>
-         )}
-         <div className={isPublic ? '' : 'border-dark-300 ml-[180px] border-l'}>{children}</div>
-      </div>
-   )
-}
