+++ src/app/layout.tsx
@@ -1,10 +1,9 @@
+import { Alert } from '@/shared/components/Alert'
+import { StoreProvider } from '@/shared/store/providers'
+import { Header } from '@/widgets/Header'
 import type { Metadata } from 'next'
 import { Inter } from 'next/font/google'
 import './globals.css'
-import { StoreProvider } from '@/shared/store/providers'
-import { Header } from '@/widgets/Header'
-import ClientRootLayout from './ClientRootLayout'
-import { Alert } from '@/shared/components/Alert'
 
 const inter = Inter({
    variable: '--font-inter',
@@ -52,7 +51,7 @@ export default function RootLayout({
          <html lang="en">
             <body className={`${inter.variable}`}>
                <Header notificationCount={4} />
-               <ClientRootLayout>{children}</ClientRootLayout>
+               {children}
                <Alert />
             </body>
          </html>
