+++ src/widgets/Sidebar/index.ts
@@ -1,3 +1,5 @@
 export * from './Sidebar'
 export { sidebarData } from './sidebarData'
 export type { SidebarItemType } from './sidebarData'
+export * from './LogoutModal'
+export * from './SidebarMenu'
